UNIX is basically a simple operating system, but you have to be a genius to understand the simplicity

+++ 0x00-hello_world/101-quote.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	char text[] = "and that piece of art is useful\" - Dora Korpar, 2015-10-19";
-	int i = 0;
+	unsigned long i = 0;
 	          
 	while(sizeof(text) > i)
 	{
